Create shared test for actsAsAMachine
